fix: quiz, problem 로직 수정 (#26)
