+++ src/main/java/com/ormee/server/service/QuizService.java
@@ -13,6 +13,7 @@ import org.springframework.stereotype.Service;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.*;
+import java.util.stream.Collectors;
 
 @Service
 public class QuizService {
@@ -29,7 +30,7 @@ public class QuizService {
         this.problemSubmitRepository = problemSubmitRepository;
     }
 
-    public void saveQuiz(UUID lectureId, QuizSaveDto quizSaveDto) {
+    public void saveQuiz(Long lectureId, QuizSaveDto quizSaveDto) {
         Lecture lecture = lectureRepository.findById(lectureId).orElseThrow(() -> new CustomException(ExceptionType.LECTURE_NOT_FOUND_EXCEPTION));
 
         Quiz quiz = Quiz.builder()
@@ -41,7 +42,6 @@ public class QuizService {
                 .openTime(quizSaveDto.getOpenTime())
                 .dueTime(quizSaveDto.getDueTime())
                 .timeLimit(quizSaveDto.getTimeLimit())
-                // 학생수 추가
                 .build();
         quiz = quizRepository.save(quiz);
 
@@ -57,20 +57,20 @@ public class QuizService {
         }
     }
 
-    public void modifyQuiz(UUID quizId, QuizSaveDto quizSaveDto) {
+    public void modifyQuiz(Long quizId, QuizSaveDto quizSaveDto) {
         Quiz quiz = quizRepository.findById(quizId).orElseThrow(() -> new CustomException(ExceptionType.QUIZ_NOT_FOUND_EXCEPTION));
         saveQuiz(quiz.getLecture().getId(), quizSaveDto);
         deleteQuiz(quizId);
     }
 
-    public List<QuizListDto> findAllByLecture(UUID lectureId, Boolean isDraft) {
+    public List<QuizListDto> findAllByLecture(Long lectureId, Boolean isDraft) {
         Lecture lecture = lectureRepository.findById(lectureId).orElseThrow(() -> new CustomException(ExceptionType.LECTURE_NOT_FOUND_EXCEPTION));
         List<Quiz> quizList = quizRepository.findAllByLectureAndIsDraftOrderByCreatedAtDesc(lecture, isDraft);
 
         return quizListToDtoList(quizList);
     }
 
-    public TeacherQuizListDto teacherQuizList(UUID lectureId, Boolean isDraft) {
+    public TeacherQuizListDto teacherQuizList(Long lectureId, Boolean isDraft) {
         Lecture lecture = lectureRepository.findById(lectureId).orElseThrow(() -> new CustomException(ExceptionType.LECTURE_NOT_FOUND_EXCEPTION));
         List<Quiz> quizList = quizRepository.findAllByLectureAndIsDraftOrderByCreatedAtDesc(lecture, isDraft);
         List<Quiz> openQuizzes = new ArrayList<>();
@@ -92,7 +92,7 @@ public class QuizService {
                 .build();
     }
 
-    public List<QuizListDto> findOpenQuizList(UUID lectureId) {
+    public List<QuizListDto> findOpenQuizList(Long lectureId) {
         Lecture lecture = lectureRepository.findById(lectureId).orElseThrow(() -> new CustomException(ExceptionType.LECTURE_NOT_FOUND_EXCEPTION));
         List<Quiz> quizList = quizRepository.findAllByLectureAndIsDraftAndIsOpenedOrderByDueTimeDesc(lecture, false, true);
 
@@ -118,19 +118,17 @@ public class QuizService {
         return quizListDtos;
     }
 
-    public void deleteQuiz(UUID quizId) {
+    public void deleteQuiz(Long quizId) {
         Quiz quiz = quizRepository.findById(quizId).orElseThrow(() -> new CustomException(ExceptionType.QUIZ_NOT_FOUND_EXCEPTION));
 
         List<Problem> problems = problemRepository.findAllByQuiz(quiz);
 
-        for(Problem problem : problems) {
-            problemRepository.delete(problem);
-        }
+        problemRepository.deleteAll(problems);
 
         quizRepository.delete(quiz);
     }
 
-    public QuizDetailDto findQuiz(UUID quizId) {
+    public QuizDetailDto findQuiz(Long quizId) {
         Quiz quiz = quizRepository.findById(quizId).orElseThrow(() -> new CustomException(ExceptionType.QUIZ_NOT_FOUND_EXCEPTION));
 
         List<Problem> problems = problemRepository.findAllByQuiz(quiz);
@@ -157,7 +155,7 @@ public class QuizService {
                 .build();
     }
 
-    public void openQuiz(UUID quizId) {
+    public void openQuiz(Long quizId) {
         Quiz quiz = quizRepository.findById(quizId).orElseThrow(() -> new CustomException(ExceptionType.QUIZ_NOT_FOUND_EXCEPTION));
         LocalDateTime now = LocalDateTime.now();
         quiz.setIsOpened(true);
@@ -165,10 +163,144 @@ public class QuizService {
         quizRepository.save(quiz);
     }
 
-    public void closeQuiz(UUID quizId) {
+    public void closeQuiz(Long quizId) {
         Quiz quiz = quizRepository.findById(quizId).orElseThrow(() -> new CustomException(ExceptionType.QUIZ_NOT_FOUND_EXCEPTION));
         LocalDateTime now = LocalDateTime.now();
         quiz.setDueTime(now);
         quizRepository.save(quiz);
+    }
+
+    public List<QuizStatsDto> getQuizStats(Long quizId) {
+        Quiz quiz = quizRepository.findById(quizId).orElseThrow(() -> new CustomException(ExceptionType.QUIZ_NOT_FOUND_EXCEPTION));
+        List<Problem> problems = problemRepository.findAllByQuiz(quiz);
+        long firstNum = problems.get(0).getId() - 1;
+
+        List<QuizStatsDto> quizStatsDtos = problems.stream()
+                .map(problem -> {
+                    List<ProblemSubmit> problemSubmits = problemSubmitRepository.findAllByProblem(problem);
+
+                    if(problemSubmits.isEmpty()) {
+                        return QuizStatsDto.builder()
+                                .problemId(problem.getId())
+                                .problemNum(problem.getId() - firstNum)
+                                .incorrectRate(0)
+                                .incorrectCount(0)
+                                .build();
+                    }
+
+                    long incorrectCount = problemSubmits.stream()
+                            .filter(problemSubmit -> !problemSubmit.getContent().equals(problem.getAnswer()))
+                            .count();
+
+                    long incorrectRate = (incorrectCount * 100) / problemSubmits.size();
+
+                    return QuizStatsDto.builder()
+                            .problemId(problem.getId())
+                            .problemNum(problem.getId() - firstNum)
+                            .incorrectRate(incorrectRate)
+                            .incorrectCount(incorrectCount)
+                            .build();
+                })
+                .sorted(Comparator
+                        .comparingLong(QuizStatsDto::getIncorrectRate).reversed()
+                        .thenComparingLong(QuizStatsDto::getIncorrectCount))
+                .collect(Collectors.toList());
+
+        for (int i = 0; i < quizStatsDtos.size(); i++) {
+            if (i == 0) {
+                quizStatsDtos.get(i).setRank(1);
+            } else {
+                QuizStatsDto current = quizStatsDtos.get(i);
+                QuizStatsDto previous = quizStatsDtos.get(i - 1);
+
+                if (current.getIncorrectRate() == previous.getIncorrectRate()) {
+                    current.setRank(previous.getRank());
+                } else {
+                    current.setRank(i + 1);
+                }
+            }
+        }
+
+        return quizStatsDtos;
+    }
+
+    public ProblemStatsDto getProblemstats(Long problemId) {
+        Problem problem = problemRepository.findById(problemId).orElseThrow(() -> new CustomException(ExceptionType.PROBLEM_NOT_FOUND_EXCEPTION));
+        List<Map<String, Object>> results = new ArrayList<>();
+
+        if (problem.getType().equals(ProblemType.CHOICE)) {
+            long noResponseCount = problemSubmitRepository.findAllByProblem(problem).stream()
+                    .map(ProblemSubmit::getContent)
+                    .filter(content -> content == null || content.isEmpty())
+                    .count();
+
+            if (noResponseCount > 0) {
+                Map<String, Object> noResponseResult = new HashMap<>();
+                noResponseResult.put("option", "무응답");
+                noResponseResult.put("count", noResponseCount);
+                results.add(0, noResponseResult);
+            }
+
+            for (String option : problem.getItems()) {
+                long count = problemSubmitRepository.countAllByProblemAndContentLike(problem, option);
+                Map<String, Object> result = new HashMap<>();
+                result.put("option", option);
+                result.put("count", count);
+                results.add(result);
+            }
+        } else {
+            List<String> submissions = problemSubmitRepository.findAllByProblem(problem)
+                    .stream()
+                    .map(ProblemSubmit::getContent)
+                    .collect(Collectors.toList());
+
+            long noResponseCount = submissions.stream()
+                    .filter(content -> content == null || content.isEmpty())
+                    .count();
+
+            submissions = submissions.stream()
+                    .filter(content -> content != null && !content.isEmpty() && !content.equals(problem.getAnswer()))
+                    .toList();
+
+            Map<String, Long> contentCounts = submissions.stream()
+                    .collect(Collectors.groupingBy(content -> content, Collectors.counting()));
+
+            List<Map.Entry<String, Long>> sortedEntries = contentCounts.entrySet().stream()
+                    .sorted((e1, e2) -> e2.getValue().compareTo(e1.getValue()))
+                    .toList();
+
+            int rank = 1;
+            int previousRank = 1;
+            long previousCount = -1;
+
+            for (Map.Entry<String, Long> entry : sortedEntries) {
+                if (previousCount != -1 && !entry.getValue().equals(previousCount)) {
+                    rank = previousRank + 1;
+                }
+                previousRank = rank;
+                previousCount = entry.getValue();
+
+                Map<String, Object> result = new HashMap<>();
+                result.put("rank", rank);
+                result.put("content", entry.getKey());
+                result.put("count", entry.getValue());
+                results.add(result);
+            }
+
+            if (noResponseCount > 0) {
+                Map<String, Object> noResponseResult = new HashMap<>();
+                noResponseResult.put("rank", -1);
+                noResponseResult.put("content", "무응답");
+                noResponseResult.put("count", noResponseCount);
+                results.add(0, noResponseResult);
+            }
+        }
+
+        return ProblemStatsDto.builder()
+                .content(problem.getContent())
+                .type(problem.getType().toString())
+                .answer(problem.getAnswer())
+                .results(results)
+                .build();
     }
 }
